Revert "#WOP-13145 (pull request #3091)"

Approved-by: Eesa Jacobs
Approved-by: Bhagyashri Uddhav Pawar

+++ app/src/main/java/za/co/woolworths/financial/services/android/models/dto/rating_n_reviews/ReviewStatistics.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class ReviewStatistics (
 
-    @SerializedName("averageRating") val averageRating : Float,
+    @SerializedName("averageRating") val averageRating : Double,
     @SerializedName("helpfulVoteCount") val helpfulVoteCount : Int,
     @SerializedName("recommendedCount") val recommendedCount : Int,
     @SerializedName("ratingDistribution") val ratingDistribution : List<RatingDistribution>,
